requirements.txt
# PatientBridge — Python dependencies

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0

# HTTP client
httpx>=0.27.0

# SMART on FHIR auth
PyJWT>=2.9.0
cryptography>=43.0

# LLM providers (grok uses httpx directly)
# openai>=1.50.0          # Uncomment when using OpenAI
# anthropic>=0.39.0       # Uncomment when using Anthropic

# OCR / PDF
pypdf>=4.0
Pillow>=10.0
pytesseract>=0.3.10
# pdf2image>=1.17.0       # Uncomment for scanned PDF OCR (requires poppler)

# MCP server
mcp[cli]>=1.0

# Dev
python-dotenv>=1.0
python-multipart>=0.0.9
